product-sale-prediction-AI/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.11
langchain-core>=0.1.9
openai>=1.3.7
streamlit>=1.32.0
httpx>=0.27.0
PyYAML>=6.0.1
jsonschema>=4.21.1
pydantic>=2.6.1
python-dotenv>=1.0.0
requests>=2.31.0
plotly>=5.18.0
pandas>=2.2.0
colorama>=0.4.6
tiktoken>=0.5.1
numpy>=1.18.0
scikit-learn>=0.24.0
xgboost>=1.3.0
joblib>=1.0.0
langchain>=0.1.0
streamlit>=1.22.0
openai>=1.10.0
httpx
requests
